+++ app/ante/evm/vesting.go
@@ -13,17 +13,12 @@
 //
 // You should have received a copy of the GNU Lesser General Public License
 // along with the Evmos packages. If not, see https://github.com/evmos/evmos/blob/main/LICENSE
-
-package ante
+package evm
 
 import (
-	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
 	errorsmod "cosmossdk.io/errors"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/cosmos-sdk/x/authz"
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	evmtypes "github.com/evmos/evmos/v11/x/evm/types"
 	vestingtypes "github.com/evmos/evmos/v11/x/vesting/types"
 )
@@ -90,104 +85,4 @@ func (vtd EthVestingTransactionDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx,
 	}
 
 	return next(ctx, tx, simulate)
-}
-
-// TODO: remove once Cosmos SDK is upgraded to v0.46
-
-// VestingDelegationDecorator validates delegation of vested coins
-type VestingDelegationDecorator struct {
-	ak  evmtypes.AccountKeeper
-	sk  vestingtypes.StakingKeeper
-	cdc codec.BinaryCodec
-}
-
-// NewVestingDelegationDecorator creates a new VestingDelegationDecorator
-func NewVestingDelegationDecorator(ak evmtypes.AccountKeeper, sk vestingtypes.StakingKeeper, cdc codec.BinaryCodec) VestingDelegationDecorator {
-	return VestingDelegationDecorator{
-		ak:  ak,
-		sk:  sk,
-		cdc: cdc,
-	}
-}
-
-// AnteHandle checks if the tx contains a staking delegation.
-// It errors if the coins are still locked or the bond amount is greater than
-// the coins already vested
-func (vdd VestingDelegationDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, next sdk.AnteHandler) (newCtx sdk.Context, err error) {
-	for _, msg := range tx.GetMsgs() {
-		switch msg := msg.(type) {
-		case *authz.MsgExec:
-			// Check for bypassing authorization
-			if err := vdd.validateAuthz(ctx, msg); err != nil {
-				return ctx, err
-			}
-		default:
-			if err := vdd.validateMsg(ctx, msg); err != nil {
-				return ctx, err
-			}
-		}
-	}
-
-	return next(ctx, tx, simulate)
-}
-
-// validateAuthz validates the authorization internal message
-func (vdd VestingDelegationDecorator) validateAuthz(ctx sdk.Context, execMsg *authz.MsgExec) error {
-	for _, v := range execMsg.Msgs {
-		var innerMsg sdk.Msg
-		if err := vdd.cdc.UnpackAny(v, &innerMsg); err != nil {
-			return errorsmod.Wrap(err, "cannot unmarshal authz exec msgs")
-		}
-
-		if err := vdd.validateMsg(ctx, innerMsg); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// validateMsg checks that the only vested coins can be delegated
-func (vdd VestingDelegationDecorator) validateMsg(ctx sdk.Context, msg sdk.Msg) error {
-	delegateMsg, ok := msg.(*stakingtypes.MsgDelegate)
-	if !ok {
-		return nil
-	}
-
-	for _, addr := range msg.GetSigners() {
-		acc := vdd.ak.GetAccount(ctx, addr)
-		if acc == nil {
-			return errorsmod.Wrapf(
-				errortypes.ErrUnknownAddress,
-				"account %s does not exist", addr,
-			)
-		}
-
-		clawbackAccount, isClawback := acc.(*vestingtypes.ClawbackVestingAccount)
-		if !isClawback {
-			// continue to next decorator as this logic only applies to vesting
-			return nil
-		}
-
-		// error if bond amount is > vested coins
-		bondDenom := vdd.sk.BondDenom(ctx)
-		coins := clawbackAccount.GetVestedOnly(ctx.BlockTime())
-		if coins == nil || coins.Empty() {
-			return errorsmod.Wrap(
-				vestingtypes.ErrInsufficientVestedCoins,
-				"account has no vested coins",
-			)
-		}
-
-		vested := coins.AmountOf(bondDenom)
-		if vested.LT(delegateMsg.Amount.Amount) {
-			return errorsmod.Wrapf(
-				vestingtypes.ErrInsufficientVestedCoins,
-				"cannot delegate unvested coins. coins vested < delegation amount (%s < %s)",
-				vested, delegateMsg.Amount.Amount,
-			)
-		}
-	}
-
-	return nil
 }
